requirements.txt
Django>=4.2
gunicorn
whitenoise
psycopg2-binary
plotly
pandas
numpy
sqlalchemy
python-decouple